Add some details in README, freeze libs, fix rubocop
rubocop adds a frozen literal instruction in every ruby file causing a
lot of noise. We're not so excited about, keeping up without and waiting
for ruby 3 to ship with frozen string literals by default :)

README details are readded from former version, that was accidentally
overwritten by the new rails init.

postgres is set as database engine and is locked to current version as
well as deivse and sqreen. A database config stub is added (untested).